requirements.txt
flask
paho-mqtt
gunicorn
flask-cors

# flask==3.1.2
# paho-mqtt==1.6.1
# gunicorn==23.0.0
